--- db/migrate/20201122005027_create_notices.rb
@@ -1,12 +1,0 @@
-class CreateNotices < ActiveRecord::Migration[6.0]
-  def change
-    create_table :notices do |t|
-      t.references :user, null: false, foreign_key: true
-      t.references :register_book, null: false, foreign_key: true
-      t.integer :noticed_flag
-
-      t.timestamps
-    end
-    add_index :notices, [:user_id, :register_book_id], unique: true
-  end
-end
