--- spec/factories/register_books.rb
@@ -1,6 +1,0 @@
-FactoryBot.define do
-  factory :register_book do
-    author_favorite { nil }
-    favored_author_book { nil }
-  end
-end
